refactor: 요약문 생성 및 저장 기능 컴포넌트화
요약문 생성 및 저장 기능 컴포넌트화 및 매뉴얼 제공
