+++ pages/demos/ten_years_ago_demo.py
@@ -1,5 +1,5 @@
 import streamlit as st
-from components.news_summary_process import run as news_summary_process
+from components.news_summary_process import select_options, show_input_news, show_summary_title
 from components.ten_years_ago_component import get_10years_ago_news, con, sidebarCon
 
 def run():
@@ -9,8 +9,15 @@ def run():
     col1, col2 = st.columns([4, 1])
 
     with col1.container():
-        news_summary_process()  # generate_summary_and_title_demo 활용
+        text, lang, length_option = select_options()
+        show_input_news(text, lang, length_option)
+
+        if st.button("Generate"):
+            if not text.strip():
+                st.warning("텍스트를 입력해주세요.")
+            else:
+                show_summary_title(text, lang)
     with col2.container(border=True):
-        con(df)     # Case 1 : 페이지 우측면에 출력 (/components/ten_years_ago_component)
+        con(df)  # Case 1 : 페이지 우측면에 출력 (/components/ten_years_ago_component)
     sidebarCon(df)  # Case 2 : Sidebar에 출력 (/components/ten_years_ago_component)
